src/requirements.txt
pandas==2.2.3
scipy==1.14.1
scikit-learn==1.5.2
zelos-demeter==0.7.4
langgraph==0.2.39
langchain-openai